set ntp server and reload config anexia-it/machine-controller@f9025b9
ANXKUBE-1356: Use Anexia NTP servers
In comparison to the official Flatcar ones, our NTP servers are not
automatically blocked by the firewall.

anexia-it/machine-controller@9710c81

ANXKUBE-1353: Apply sysctls automatically with kubelet
sysctl --system can be executed multiple times, so this should be a safe
change. It's also just a soft dependency (Wants instead of Requires), so
even in case of an error, it would not block the startup of kubelet.

It's still unclear why the unit is sometimes not executed and right now,
I assume it has to do something with the network dependences.Please enter the commit message for your changes. Lines starting